requirements.txt
Flask==1.1.2
google-cloud-datastore==1.15.0
google-auth==1.21.2
requests==2.24.0
google-cloud-storage==1.17.0
